+++ local/src/main/java/com/cairosquad/local/search/discovery/DiscoveryDataSourceImpl.kt
@@ -1,11 +1,11 @@
 package com.cairosquad.local.search.discovery
 
 import com.cairosquad.local.search.discovery.dao.DiscoveryDao
-import com.cairosquad.repository.common.mappers.toPersonalizedMoviesIdsDto
-import com.cairosquad.repository.common.mappers.toSuggestedMoviesIds
 import com.cairosquad.repository.search.data_source.local.CacheDataSource
 import com.cairosquad.repository.search.data_source.local.DiscoveryDataSource
 import com.cairosquad.repository.search.data_source.local.dto.MovieCacheDto
+import com.cairosquad.repository.utils.mappers.toPersonalizedMoviesIdsDto
+import com.cairosquad.repository.utils.mappers.toSuggestedMoviesIds
 
 class DiscoveryDataSourceImpl(
     private val cacheDataSource: CacheDataSource,
