+++ remote/src/main/java/com/cairosquad/remote/common/utils/callApi.kt
@@ -1,6 +1,6 @@
 package com.cairosquad.remote.common.utils
 
-import com.cairosquad.repository.common.exception.NoInternetException
+import com.cairosquad.repository.utils.exception.NoInternetException
 import io.ktor.client.statement.HttpResponse
 import java.io.IOException
