+++ repository/src/main/java/com/cairosquad/repository/search/MoviesRepositoryImpl.kt
@@ -4,12 +4,12 @@ import com.cairosquad.domain.repository.MoviesRepository
 import com.cairosquad.entity.Artist
 import com.cairosquad.entity.Movie
 import com.cairosquad.entity.Review
-import com.cairosquad.repository.common.mappers.tryToCall
 import com.cairosquad.repository.search.data_source.local.DiscoveryDataSource
 import com.cairosquad.repository.search.data_source.local.dto.toCacheDto
 import com.cairosquad.repository.search.data_source.local.dto.toEntity
 import com.cairosquad.repository.search.data_source.remote.RemoteMovieDiscoveryDataSource
 import com.cairosquad.repository.search.data_source.remote.dto.toEntity
+import com.cairosquad.repository.utils.mappers.tryToCall
 import java.util.Date
 
 class MoviesRepositoryImpl(
